Moved NewJobDataToAdd to schedule.tsx
