+++ src/react/components/AddJobFormModel.tsx
@@ -1,14 +1,8 @@
 // AddJobFormModal.tsx
 import React, { useState, useEffect } from 'react';
+import type { NewJobDataForAdd } from '../pages/Schedule'
 
 // Define NewJobDataForAdd interface directly in this file
-export interface NewJobDataForAdd {
-    invoiceId: String;
-    client: String;
-    name: String;
-    type: String; // Added 'type' as per your mock data
-    due: Date;
-}
 
 export interface AddJobFormModelProps {
     isOpen: boolean;
