+++ src/react/components/JobTable.tsx
@@ -1,7 +1,7 @@
 import { useState, useEffect } from 'react';
 import AddSubJobFormModal from "../components/AddSubJobFormModal";
 import JobTableRow from './JobTableRow';
-import type {NewSubJobDataForAdd} from '../pages/Schedule'
+import type { NewSubJobDataForAdd } from '../pages/Schedule'
